backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
openai
python-dotenv
email-validator
passlib[bcrypt]
python-jose[cryptography]httpx
httpx